pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "themis"
version = "0.2.0"
description = "Compliance and risk management platform"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "Themis Team", email = "team@themis.local"}
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "bcrypt",
    "itsdangerous",
    "jinja2",
    "python-multipart",
    "pydantic-settings",
    "python-dotenv",
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "black==24.1.1",
    "ruff==0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.ruff]
line-length = 100
target-version = "py314"
select = ["E", "F", "W"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
